requirements.txt
fastapi
uvicorn
gradio-client
pillow
cachetools
python-multipart
jinja2
retry
slowapi